Added ability to hook as an API
I added the ability to use this as an api to translate through other plugins by using the 'McTranslate.java' class. By using the methods in there I can save translations to a file for later use if they're only needed to be grabbed once, otherwise you can translate real time for uses such as talking in game or on items that are created on runtime. and aren't saved to a file. With the saving to the file ability the path of the string can be set to be recieved on a later date. This takes a load of work that would have to be done each time a plugin wanted to multi-version, and this makes it so much easier to do that.